:hammer: Refactor metrics folder
