+++ torch_uncertainty/metrics/classification/brier_score.py
@@ -8,7 +8,7 @@ from torchmetrics.utilities.data import dim_zero_cat
 
 
 class BrierScore(Metric):
-    is_differentiable: bool = False
+    is_differentiable: bool = True
     higher_is_better: bool | None = False
     full_state_update: bool = False
