+++ torch_uncertainty/metrics/classification/disagreement.py
