+++ torch_uncertainty/metrics/classification/mutual_information.py
