+++ torch_uncertainty/metrics/classification/variation_ratio.py
